requirements.txt
torch
torchvision
streamlit
Pillow
numpy
opencv-python
